--- src/main/java/com/sparta/twotwo/auth/service/MemberDetailService.java
@@ -1,70 +1,0 @@
-package com.sparta.twotwo.auth.service;
-
-import com.sparta.twotwo.auth.util.AuthorityUtil;
-import com.sparta.twotwo.members.entity.Member;
-import com.sparta.twotwo.members.repository.MemberRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-import java.util.Collection;
-import java.util.Optional;
-
-@Service
-@RequiredArgsConstructor
-public class MemberDetailService implements UserDetailsService {
-
-    private final MemberRepository memberRepository;
-    private final AuthorityUtil authorityUtil;
-
-
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        Optional<Member> member = memberRepository.findByUsername(username);
-        Member findMember = member.orElseThrow(() -> new UsernameNotFoundException("Username not found"));
-        System.out.println(username);
-        System.out.println(member);
-        System.out.println(findMember.getPassword());
-
-        return new MemberDetail(findMember);
-    }
-
-    private final class MemberDetail extends Member implements UserDetails {
-        MemberDetail (Member member) {
-            setMember_id(member.getMember_id());
-            setUsername(member.getUsername());
-            setNickname(member.getNickname());
-            setEmail(member.getEmail());
-            setPassword(member.getPassword());
-            setRoles(member.getRoles());
-        }
-
-        @Override
-        public Collection<? extends GrantedAuthority> getAuthorities() {
-            return authorityUtil.createAuthorities(this.getRoles());
-        }
-
-        @Override
-        public boolean isAccountNonExpired() {
-            return UserDetails.super.isAccountNonExpired();
-        }
-
-        @Override
-        public boolean isAccountNonLocked() {
-            return UserDetails.super.isAccountNonLocked();
-        }
-
-        @Override
-        public boolean isCredentialsNonExpired() {
-            return UserDetails.super.isCredentialsNonExpired();
-        }
-
-        @Override
-        public boolean isEnabled() {
-            return UserDetails.super.isEnabled();
-        }
-    }
-}
