+++ src/main/java/com/sparta/twotwo/auth/service/MemberDetailsService.java
@@ -1,0 +1,32 @@
+package com.sparta.twotwo.auth.service;
+
+import com.sparta.twotwo.auth.util.AuthorityUtil;
+import com.sparta.twotwo.members.entity.Member;
+import com.sparta.twotwo.members.repository.MemberRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+import java.util.Collection;
+import java.util.Optional;
+
+@Service
+@RequiredArgsConstructor
+public class MemberDetailsService implements UserDetailsService {
+
+    private final MemberRepository memberRepository;
+
+    @Override
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        Optional<Member> member = memberRepository.findByUsername(username);
+        Member findMember = member.orElseThrow(() -> new UsernameNotFoundException("Username not found"));
+        System.out.println(username);
+        System.out.println(member);
+        System.out.println(findMember.getPassword());
+
+        return new MemberDetails(findMember);
+    }
+}
